+++ test/Savvyio.Extensions.Newtonsoft.Json.Tests/Commands/Messaging/Cryptography/MessageExtensionsTest.cs
@@ -6,8 +6,8 @@ using Cuemon.Extensions.IO;
 using Cuemon.Extensions.Xunit;
 using Savvyio.Assets.Commands;
 using Savvyio.Commands.Messaging;
-using Savvyio.Commands.Messaging.Cryptography;
 using Savvyio.Messaging;
+using Savvyio.Messaging.Cryptography;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -30,62 +30,29 @@ namespace Savvyio.Extensions.Newtonsoft.Json.Commands.Messaging.Cryptography
                 o.Time = utc;
             });
 
-            var sut3 = sut2.Sign(new NewtonsoftJsonMarshaller(), o =>
+            var sut3 = sut2.Sign(NewtonsoftJsonMarshaller.Default, o =>
             {
                 o.SignatureSecret = new byte[] { 1, 2, 3 };
             });
 
-            var json = new NewtonsoftJsonMarshaller().Serialize(sut3);
+            var json = NewtonsoftJsonMarshaller.Default.Serialize(sut3);
             var jsonString = json.ToEncodedString(o => o.LeaveOpen = true);
 
             TestOutput.WriteLine(jsonString);
 
-            var sut4 = new NewtonsoftJsonMarshaller().Deserialize<IMessage<CreateMemberCommand>>(json);
+            sut3.CheckSignature(NewtonsoftJsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 1, 2, 3 });
+
+            var signatureException = Assert.Throws<ArgumentOutOfRangeException>(() =>
+            {
+                sut3.CheckSignature(NewtonsoftJsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 3, 2, 1 });
+            });
+            Assert.StartsWith("The signature of the message does not match the cryptographically calculated value. Either you are using an incorrect secret and/or algorithm or the message has been tampered with.", signatureException.Message);
+
+            var sut4 = NewtonsoftJsonMarshaller.Default.Deserialize<IMessage<CreateMemberCommand>>(json);
 
             Assert.Equivalent(sut2, sut4, true);
 
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "CreateMemberCommand",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "age": 21,
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.Commands.CreateMemberCommand, Savvyio.Assets.Tests",
-                                   "correlationId": "3eefdef050c340bfba100bd49c58c181"
-                                 }
-                               },
-                               "signature": "194374d817b2b3f9f8ea1ce6557bbb4a96200de2a759270df96d2a20cb28cdae"
-                             }
-                             """.ReplaceLineEndings(), jsonString);
-            }
-            else
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "CreateMemberCommand",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "age": 21,
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.Commands.CreateMemberCommand, Savvyio.Assets.Tests",
-                                   "correlationId": "3eefdef050c340bfba100bd49c58c181"
-                                 }
-                               },
-                               "signature": "8d15d927f35b57f8e89392952e43e5cca32d240902c01bc00baaa819f71c5a5a"
-                             }
-                             """, jsonString);
-            }
+            Assert.Equal("""{"id":"2d4030d32a254ee8a27046e5bafe696a","source":"https://fancy.api/members","type":"CreateMemberCommand","time":"2023-11-16T23:24:17.8414532Z","data":{"name":"Jane Doe","age":21,"emailAddress":"jd@office.com","metadata":{"memberType":"Savvyio.Assets.Commands.CreateMemberCommand, Savvyio.Assets.Tests","correlationId":"3eefdef050c340bfba100bd49c58c181"}},"signature":"0cb0f1b239d9a31fad1dcbf9819e105d841f860c818f389d487d0693ed014c5b"}""", jsonString);
 
         }
     }