+++ test/Savvyio.Extensions.Text.Json.Tests/EventDriven/Messaging/Cryptography/IntegrationEventExtensionsTest.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.Globalization;
-using System.Runtime.InteropServices;
 using Cuemon.Extensions;
 using Cuemon.Extensions.IO;
-using Cuemon.Extensions.Text.Json.Formatters;
 using Cuemon.Extensions.Xunit;
 using Savvyio.Assets.EventDriven;
 using Savvyio.EventDriven.Messaging;
-using Savvyio.EventDriven.Messaging.Cryptography;
 using Savvyio.Messaging.Cryptography;
 using Xunit;
 using Xunit.Abstractions;
@@ -29,61 +26,28 @@ namespace Savvyio.Extensions.Text.Json.EventDriven.Messaging.Cryptography
             {
                 o.MessageId = "2d4030d32a254ee8a27046e5bafe696a";
                 o.Time = utc;
-            }).Sign(new JsonMarshaller(), o =>
+            }).Sign(JsonMarshaller.Default, o =>
             {
                 o.SignatureSecret = new byte[] { 1, 2, 3 };
             });
-            var json = new JsonMarshaller().Serialize(sut2);
+            var json = JsonMarshaller.Default.Serialize(sut2);
             var jsonString = json.ToEncodedString(o => o.LeaveOpen = true);
 
             TestOutput.WriteLine(jsonString);
 
-            var sut4 = new JsonMarshaller().Deserialize<ISignedMessage<MemberCreated>>(json);
+            sut2.CheckSignature(JsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 1, 2, 3 });
+
+            var signatureException = Assert.Throws<ArgumentOutOfRangeException>(() =>
+            {
+                sut2.CheckSignature(JsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 3, 2, 1 });
+            });
+            Assert.StartsWith("The signature of the message does not match the cryptographically calculated value. Either you are using an incorrect secret and/or algorithm or the message has been tampered with.", signatureException.Message);
+
+            var sut4 = JsonMarshaller.Default.Deserialize<ISignedMessage<MemberCreated>>(json);
 
             Assert.Equivalent(sut2, sut4, true);
 
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "MemberCreated",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests",
-                                   "eventId": "69bccf3b1117425397c5ed9ed757bb0f",
-                                   "timestamp": "2023-11-16T23:24:17.8414532Z"
-                                 }
-                               },
-                               "signature": "ebad495fa4b3451f043ecc859a66fb2466ac3d283360e7017d93d17ef6cbbab0"
-                             }
-                             """.ReplaceLineEndings(), jsonString);
-            }
-            else
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "MemberCreated",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests",
-                                   "eventId": "69bccf3b1117425397c5ed9ed757bb0f",
-                                   "timestamp": "2023-11-16T23:24:17.8414532Z"
-                                 }
-                               },
-                               "signature": "e41be47f5cf809d3f3d6b8ea7a35ed46c9d1f6c4e7b2178cc387990eed87b671"
-                             }
-                             """.ReplaceLineEndings(), jsonString);
-            }
+            Assert.Equal("""{"id":"2d4030d32a254ee8a27046e5bafe696a","source":"https://fancy.api/members","type":"MemberCreated","time":"2023-11-16T23:24:17.8414532Z","data":{"name":"Jane Doe","emailAddress":"jd@office.com","metadata":{"memberType":"Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests","eventId":"69bccf3b1117425397c5ed9ed757bb0f","timestamp":"2023-11-16T23:24:17.8414532Z"}},"signature":"e0b6ca2328019e236d9cd5590c20d12a2a92797f8feb34c044b92f36df013fac"}""", jsonString);
         }
 
         [Fact]
@@ -95,61 +59,28 @@ namespace Savvyio.Extensions.Text.Json.EventDriven.Messaging.Cryptography
             {
                 o.MessageId = "2d4030d32a254ee8a27046e5bafe696a";
                 o.Time = utc;
-            }).Sign(new JsonMarshaller(), o =>
+            }).Sign(JsonMarshaller.Default, o =>
             {
                 o.SignatureSecret = new byte[] { 1, 2, 3 };
             });
-            var json = JsonFormatter.SerializeObject(sut2);
+            var json = JsonMarshaller.Default.Serialize(sut2);
             var jsonString = json.ToEncodedString(o => o.LeaveOpen = true);
 
             TestOutput.WriteLine(jsonString);
 
-            var sut4 = new JsonMarshaller().Deserialize<SignedMessage<MemberCreated>>(json);
+            sut2.CheckSignature(JsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 1, 2, 3 });
+
+            var signatureException = Assert.Throws<ArgumentOutOfRangeException>(() =>
+            {
+                sut2.CheckSignature(JsonMarshaller.Default, o => o.SignatureSecret = new byte[] { 3, 2, 1 });
+            });
+            Assert.StartsWith("The signature of the message does not match the cryptographically calculated value. Either you are using an incorrect secret and/or algorithm or the message has been tampered with.", signatureException.Message);
+
+            var sut4 = JsonMarshaller.Default.Deserialize<SignedMessage<MemberCreated>>(json);
 
             Assert.Equivalent(sut2, sut4, true);
 
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "MemberCreated",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests",
-                                   "eventId": "69bccf3b1117425397c5ed9ed757bb0f",
-                                   "timestamp": "2023-11-16T23:24:17.8414532Z"
-                                 }
-                               },
-                               "signature": "ebad495fa4b3451f043ecc859a66fb2466ac3d283360e7017d93d17ef6cbbab0"
-                             }
-                             """.ReplaceLineEndings(), jsonString);
-            }
-            else
-            {
-                Assert.Equal("""
-                             {
-                               "id": "2d4030d32a254ee8a27046e5bafe696a",
-                               "source": "https://fancy.api/members",
-                               "type": "MemberCreated",
-                               "time": "2023-11-16T23:24:17.8414532Z",
-                               "data": {
-                                 "name": "Jane Doe",
-                                 "emailAddress": "jd@office.com",
-                                 "metadata": {
-                                   "memberType": "Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests",
-                                   "eventId": "69bccf3b1117425397c5ed9ed757bb0f",
-                                   "timestamp": "2023-11-16T23:24:17.8414532Z"
-                                 }
-                               },
-                               "signature": "e41be47f5cf809d3f3d6b8ea7a35ed46c9d1f6c4e7b2178cc387990eed87b671"
-                             }
-                             """.ReplaceLineEndings(), jsonString);
-            }
+            Assert.Equal("""{"id":"2d4030d32a254ee8a27046e5bafe696a","source":"https://fancy.api/members","type":"MemberCreated","time":"2023-11-16T23:24:17.8414532Z","data":{"name":"Jane Doe","emailAddress":"jd@office.com","metadata":{"memberType":"Savvyio.Assets.EventDriven.MemberCreated, Savvyio.Assets.Tests","eventId":"69bccf3b1117425397c5ed9ed757bb0f","timestamp":"2023-11-16T23:24:17.8414532Z"}},"signature":"e0b6ca2328019e236d9cd5590c20d12a2a92797f8feb34c044b92f36df013fac"}""", jsonString);
         }
     }
 }